% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{add_random_intercept}
\alias{add_random_intercept}
\title{Add Random Intercept}
\usage{
add_random_intercept(.data, variance = 1)
}
\arguments{
\item{.data}{A dataframe containing the design structure}

\item{variance}{The variance of the random intercept}
}
\value{
A dataframe with the random intercept added
}
\description{
This function adds a random intercept to the design dataframe.
}
\examples{
design_df <- data.frame(cluster = rep(1:5, each = 10))
add_random_intercept(design_df, variance = 1)
}
