% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_site_data}
\alias{generate_site_data}
\title{Generate Example Site Data}
\usage{
generate_site_data(n, min_pcps, max_pcps)
}
\arguments{
\item{n}{Number of sites to generate}

\item{min_pcps}{Minimum number of primary care providers per site}

\item{max_pcps}{Maximum number of primary care providers per site}
}
\value{
A data frame with site information
}
\description{
This function generates example site data for use in stepped wedge trial simulations.
}
\examples{
generate_site_data(10, 1, 5)
}
