% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{generate_stepped_wedge_schedule}
\alias{generate_stepped_wedge_schedule}
\title{Generate Stepped Wedge Schedule}
\usage{
generate_stepped_wedge_schedule(n_sites, n_timepoints, n_baseline_timepoints)
}
\arguments{
\item{n_sites}{Number of sites}

\item{n_timepoints}{Number of time points}

\item{n_baseline_timepoints}{Number of time points where all sites are in control condition}
}
\value{
A data frame with the stepped wedge schedule
}
\description{
This function generates a stepped wedge schedule for a given number of sites and time points.
}
\examples{
generate_stepped_wedge_schedule(10, 9, 2)
}
