% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{join_unit_info}
\alias{join_unit_info}
\title{Join Unit Information to Schedule}
\usage{
join_unit_info(
  long_schedule,
  unit_info,
  by = NULL,
  uncount_unit = NULL,
  .id = NULL
)
}
\arguments{
\item{long_schedule}{A data frame containing the long-format schedule.}

\item{unit_info}{A data frame containing unit information.}

\item{by}{Column(s) to join by.}

\item{uncount_unit}{Column to use for uncounting (expanding) rows.}

\item{.id}{Name of the new ID column created when uncounting.}
}
\value{
A data frame with unit information joined to the schedule.
}
\description{
Join Unit Information to Schedule
}
\examples{
long_schedule <- data.frame(
  cohort = c("A", "A", "B", "B"),
  time = c(1, 2, 1, 2),
  condition = c("control", "intervention", "control", "control")
)
unit_info <- data.frame(
  cohort = c("A", "B"),
  num_units = c(2, 3)
)
join_unit_info(long_schedule, unit_info, by = "cohort", uncount_unit = num_units)
}
