% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{multiply}
\alias{multiply}
\title{Title}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{the first number}

\item{b}{the second number}
}
\value{
the product
}
\description{
Title
}
\examples{
multiply(2, 2)
}
