% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{read_excel_example}
\alias{read_excel_example}
\title{Read Example Excel Sheet}
\usage{
read_excel_example(example = 1)
}
\arguments{
\item{example}{Sheet number to read from the example Excel file.}

\item{file_name}{Name of the Excel file to read from (default: "common_rollout_designs.xlsx").}
}
\value{
A data frame containing the data from the specified Excel sheet.
}
\description{
Read Example Excel Sheet
}
\examples{
# Read the first sheet from the default example Excel file
read_excel_example(1)

# Read the second sheet from a specific example Excel file
read_excel_example(2, "another_example_file.xlsx")
}
