% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollout-package.R
\docType{package}
\name{rollout-package}
\alias{rollout}
\alias{rollout-package}
\title{rollout: Tools for Designing, Simulating, and Analyzing Implementation Rollout Trials}
\description{
Provides a unified framework for designing, simulating, and analyzing implementation rollout trials, including stepped wedge, sequential rollout, head-to-head, multi-condition, and rollout implementation optimization designs. The package enables users to flexibly specify rollout schedules, incorporate site-level and nested data structures, generate outcomes under rich hierarchical models, and evaluate analytic strategies through simulation-based power analysis. By separating data generation from model fitting, the tools support assessment of bias, Type I error, and robustness to model misspecification. The workflow integrates with standard mixed-effects modeling approaches and the tidyverse ecosystem, offering transparent and reproducible tools for implementation scientists and applied statisticians.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iancero/rollout}
  \item \url{https://iancero.github.io/rollout/}
  \item Report bugs at \url{https://github.com/iancero/rollout/issues}
}

}
\author{
\strong{Maintainer}: Ian Cero \email{ian_cero@urmc.rochester.edu} (\href{https://orcid.org/0000-0002-2862-0450}{ORCID})

Authors:
\itemize{
  \item Ian Cero \email{ian_cero@urmc.rochester.edu} (\href{https://orcid.org/0000-0002-2862-0450}{ORCID})
  \item C. Hendricks Brown \email{hendricks.brown@northwestern.edu} (\href{https://orcid.org/0000-0002-0294-2419}{ORCID})
}

}
\keyword{internal}
