% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{sim_normal_error}
\alias{sim_normal_error}
\title{Add normally-distributed error to a dataset}
\usage{
sim_normal_error(.data, sd = 1)
}
\arguments{
\item{.data}{A dataframe}

\item{sd}{The standard deviation of the error term}
}
\value{
The original \code{.data} dataframe, with a new column called
\code{.error} appended to it.
}
\description{
This function adds a column of normally distributed error to a dataset and is
designed for the simulation of data for a rollout design.
}
\details{
This function is intended to be used during the simulation of a rollout-based
dataset. The error term is normally distributed with a mean of 0 and a standard
deviation of 1 (by default).
}
\examples{
mtcars |>
 sim_normal_error()

mtcars |>
 sim_normal_error(sd = 2)

}
