% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_random_unit}
\alias{simulate_random_unit}
\title{Simulate Random Unit}
\usage{
simulate_random_unit(design_df, ...)
}
\arguments{
\item{design_df}{A dataframe containing the design structure}

\item{...}{A single named argument specifying the variable to simulate and its distribution}
}
\value{
A dataframe with the new simulated variable added
}
\description{
This function simulates a random variable for each unit in the design dataframe.
}
\examples{
design_df <- data.frame(cluster = rep(1:5, each = 10))
simulate_random_unit(design_df, patients = rpois(n(), lambda = 5))
}
