% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_data_ex}
\alias{site_data_ex}
\title{Example Site Data for Stepped Wedge Trial}
\format{
A data frame with rows and columns:
\describe{
\item{cohort}{Factor or character, unique identifier for each site/cluster}
\item{num_pcps}{Numeric, number of primary care providers at each site}
\item{...}{Add other relevant variables that are in the site_data_ex dataset}
}
}
\source{
Simulated data for demonstration purposes
}
\usage{
site_data_ex
}
\description{
A dataset containing information about sites (clusters) for a stepped wedge trial.
}
\keyword{datasets}
