% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{add_error}
\alias{add_error}
\title{Add Error Term}
\usage{
add_error(.data, variance = 1)
}
\arguments{
\item{.data}{A dataframe containing the design structure}

\item{variance}{The variance of the error term}
}
\value{
A dataframe with the error term added
}
\description{
This function adds a random error term to the design dataframe.
}
\examples{
design_df <- data.frame(cluster = rep(1:5, each = 10))
add_error(design_df, variance = 0.5)
}
