% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{add_fixed_effect}
\alias{add_fixed_effect}
\title{Add Fixed Effect}
\usage{
add_fixed_effect(design_df, ...)
}
\arguments{
\item{design_df}{A dataframe containing the design structure}

\item{...}{A single named argument specifying the fixed effect to add}
}
\value{
A dataframe with the new fixed effect added
}
\description{
This function adds a fixed effect to the design dataframe.
}
\examples{
design_df <- data.frame(time = 1:10)
add_fixed_effect(design_df, time_effect = 0.1 * time)
}
