% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{add_random_effect}
\alias{add_random_effect}
\title{Add Random Effect}
\usage{
add_random_effect(design_df, ..., nesting_strc)
}
\arguments{
\item{design_df}{A dataframe containing the design structure}

\item{...}{A single named argument specifying the random effect to add}

\item{nesting_strc}{The nesting structure for the random effect}
}
\value{
A dataframe with the new random effect added
}
\description{
This function adds a random effect to the design dataframe.
}
\examples{
design_df <- data.frame(cluster = rep(1:5, each = 10), time = rep(1:10, 5))
add_random_effect(design_df, time_effect = rnorm(n(), sd = 0.1), nesting_strc = "cluster")
}
