% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{pivot_schedule_longer}
\alias{pivot_schedule_longer}
\title{Pivot Schedule to Longer Format}
\usage{
pivot_schedule_longer(
  schedule,
  time_cols,
  names_to = "time",
  names_pattern = ".*(\\\\d+)",
  names_transform = as.numeric,
  values_to = "condition",
  values_transform = as.factor,
  cohort_name = cohort,
  condition_time = TRUE,
  condition_time_int = 0
)
}
\arguments{
\item{schedule}{A data frame containing the schedule.}

\item{time_cols}{Columns to pivot.}

\item{names_to}{Name of the new 'names' column.}

\item{names_pattern}{Regex pattern for extracting time from column names.}

\item{names_transform}{Function to transform the time values.}

\item{values_to}{Name of the new 'values' column.}

\item{values_transform}{Function to transform the condition values.}

\item{cohort_name}{Name of the cohort column.}

\item{condition_time}{Whether to calculate condition time.}

\item{condition_time_int}{Initial value for condition time.}
}
\value{
A data frame in long format.
}
\description{
Pivot Schedule to Longer Format
}
\examples{
schedule <- data.frame(
  cohort = c("A", "B"),
  t1 = c("control", "control"),
  t2 = c("intervention", "control"),
  t3 = c("intervention", "intervention")
)
pivot_schedule_longer(schedule, t1:t3)
}
